+++ src/Transports/Types.ts
@@ -30,13 +30,6 @@ export interface CastoeFileOptions {
 export interface CastoeFileStreamOptions {
   flags?: string;
   encoding?: string;
-  fd?: number;
-  mode: number;
-  autoClose: boolean;
-  emitClose: boolean;
-  start: number;
-  end: number;
-  highWaterMark: number;
 }
 
 export interface CastoeStreamOptions {